Refactor Doctor Model to create off of DoctorFormatter
